self_hosted/requirements.txt
flask==2.3.3
gunicorn==21.2.0
eventlet==0.34.2
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-login==0.6.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
discord.py==2.3.2
alpaca-trade-api==3.0.2
pandas==2.1.1
numpy==1.26.0
python-dateutil==2.8.2
apscheduler==3.10.4
Werkzeug==2.3.7
email-validator==2.1.0
"""Optional AI components"""
# anthropic==0.7.4  # Uncomment if using Anthropic Claude
# openai==1.3.7    # Uncomment if using OpenAI GPT models